+++ playwright/surveys/feedback-widget.spec.ts
@@ -39,7 +39,7 @@ const appearanceWithThanks = {
     thankyouMessageBody: 'We appreciate your feedback.',
 }
 
-const black = 'rgb(0, 0, 0)'
+const black = 'rgb(2, 6, 23)'
 const white = 'rgb(255, 255, 255)'
 
 test.describe('surveys - feedback widget', () => {
@@ -300,6 +300,7 @@ test.describe('surveys - feedback widget', () => {
         // 5. Close the thank you message and click the survey tab again
         await page.locator('.PostHogSurvey-123').locator('.form-submit').click()
         await expect(page.locator('.PostHogSurvey-123').locator('.survey-form')).not.toBeVisible()
+        await page.waitForTimeout(300)
 
         // Open the survey again
         await page.locator('.PostHogSurvey-123').locator('.ph-survey-widget-tab').click()