+++ src/__tests__/posthog-core.test.ts
@@ -53,7 +53,7 @@ describe('posthog core', () => {
             })
 
             // act
-            const actual = posthog._calculate_event_properties(eventName, eventProperties, new Date())
+            const actual = posthog.calculateEventProperties(eventName, eventProperties, new Date())
 
             // assert
             expect(actual['event']).toBe('prop')